Moved hour/min formatter to helper file
